Just used an open source GTAO...

+++ include/Core/Timer.hpp
@@ -15,7 +15,7 @@ public:
     void Reset();
 
     float GetElapsedSeconds();
-    long GetElapsedMilliseconds();
+    float GetElapsedMilliseconds();
 
 private:
     std::chrono::high_resolution_clock::time_point start;
